Refactor Example/Where/Zone in their own module example.py (api break)
